+++ no_proto_rs/src/pointer/bool.rs
@@ -30,12 +30,22 @@ use alloc::string::ToString;
 use super::NP_Cursor;
 
 impl<'value> super::NP_Scalar<'value> for bool {
+
     fn schema_default(_schema: &NP_Parsed_Schema) -> Option<Self> where Self: Sized {
         Some(Self::default())
     }
+
 }
 
 impl<'value> NP_Value<'value> for bool {
+
+    fn np_max_value<M: NP_Memory>(_cursor: &NP_Cursor, _memory: &M) -> Option<Self> {
+        Some(true)
+    }
+
+    fn np_min_value<M: NP_Memory>(_cursor: &NP_Cursor, _memory: &M) -> Option<Self> {
+        Some(false)
+    }
 
     fn type_idx() -> (&'value str, NP_TypeKeys) { ("bool", NP_TypeKeys::Boolean) }
     fn self_type_idx(&self) -> (&'value str, NP_TypeKeys) { ("bool", NP_TypeKeys::Boolean) }