+++ no_proto_rs/src/pointer/string.rs
@@ -62,6 +62,49 @@ impl<'value> NP_Scalar<'value> for String {
 
 
 impl<'value> NP_Value<'value> for String {
+
+    fn np_max_value<M: NP_Memory>(cursor: &NP_Cursor, memory: &M) -> Option<Self> {
+        let size = match memory.get_schema(cursor.schema_addr) {
+            NP_Parsed_Schema::UTF8String { size, .. } => {
+                *size
+            },
+            _ => 0
+        };
+
+        if size == 0 {
+            None
+        } else {
+            let mut value: String = String::with_capacity(size as usize);
+
+            for _x in 0..size {
+                value.push_str(unsafe { str::from_utf8_unchecked(&[128])});
+            }
+
+            Some(value)
+        }
+    }
+
+    fn np_min_value<M: NP_Memory>(cursor: &NP_Cursor, memory: &M) -> Option<Self> {
+        let size = match memory.get_schema(cursor.schema_addr) {
+            NP_Parsed_Schema::UTF8String { size, .. } => {
+                *size
+            },
+            _ => 0
+        };
+
+        if size == 0 {
+            None
+        } else {
+            let mut value: String = String::with_capacity(size as usize);
+
+            for _x in 0..size {
+                value.push_str(unsafe { str::from_utf8_unchecked(&[0])});
+            }
+
+            Some(value)
+        }
+    }
+
     fn type_idx() -> (&'value str, NP_TypeKeys) {
         ("string", NP_TypeKeys::UTF8String)
     }
@@ -304,9 +347,19 @@ impl<'value> NP_Scalar<'value> for NP_String<'value> {
     fn schema_default(_schema: &NP_Parsed_Schema) -> Option<Self> where Self: Sized {
         None
     }
+
 }
 
 impl<'value> NP_Value<'value> for NP_String<'value> {
+
+    fn np_max_value<M: NP_Memory>(_cursor: &NP_Cursor, _memory: &M) -> Option<Self> {
+        None
+    }
+
+    fn np_min_value<M: NP_Memory>(_cursor: &NP_Cursor, _memory: &M) -> Option<Self> {
+        None
+    }
+
     fn type_idx() -> (&'value str, NP_TypeKeys) { String::type_idx() }
     fn self_type_idx(&self) -> (&'value str, NP_TypeKeys) { String::default().self_type_idx() }
 
@@ -316,7 +369,7 @@ impl<'value> NP_Value<'value> for NP_String<'value> {
 
     fn set_value<'set, M: NP_Memory>(cursor: NP_Cursor, memory: &'set M, value: Self) -> Result<NP_Cursor, NP_Error> where Self: 'set + Sized {
 
-        let c_value = cursor.get_value(memory);
+        let c_value = || { cursor.get_value(memory) };
 
         let (size, case) = match memory.get_schema(cursor.schema_addr) {
             NP_Parsed_Schema::UTF8String { size, case, .. } => (*size, *case),
@@ -347,7 +400,7 @@ impl<'value> NP_Value<'value> for NP_String<'value> {
         if size > 0 {
             // fixed size bytes
     
-            if c_value.get_addr_value() == 0 {
+            if c_value().get_addr_value() == 0 {
                 // malloc new bytes
     
                 let mut empty_bytes: Vec<u8> = Vec::with_capacity(size as usize);
@@ -356,10 +409,10 @@ impl<'value> NP_Value<'value> for NP_String<'value> {
                 }
     
                 let new_addr = memory.malloc(empty_bytes)? as usize;
-                c_value.set_addr_value(new_addr as u16);
+                c_value().set_addr_value(new_addr as u16);
             }
 
-            let addr = c_value.get_addr_value() as usize;
+            let addr = c_value().get_addr_value() as usize;
             write_bytes = memory.write_bytes();
     
             for x in 0..(size as usize) {
@@ -376,7 +429,7 @@ impl<'value> NP_Value<'value> for NP_String<'value> {
         }
     
         // flexible size
-        let addr_value = c_value.get_addr_value() as usize;
+        let addr_value = c_value().get_addr_value() as usize;
     
         let prev_size: usize = if addr_value != 0 {
             let size_bytes: &[u8; 2] = memory.get_2_bytes(addr_value).unwrap_or(&[0; 2]);
@@ -418,7 +471,7 @@ impl<'value> NP_Value<'value> for NP_String<'value> {
                 memory.malloc_borrow(&size_bytes)?
             };
     
-            c_value.set_addr_value(new_addr as u16);
+            c_value().set_addr_value(new_addr as u16);
     
             memory.malloc_borrow(bytes)?;