+++ no_proto_rs/src/pointer/uuid.rs
@@ -151,6 +151,14 @@ impl<'value> NP_Scalar<'value> for NP_UUID {
 
 impl<'value> NP_Value<'value> for NP_UUID {
 
+    fn np_max_value<M: NP_Memory>(_cursor: &NP_Cursor, _memory: &M) -> Option<Self> {
+        Some(NP_UUID { value: [255u8; 16]})
+    }
+
+    fn np_min_value<M: NP_Memory>(_cursor: &NP_Cursor, _memory: &M) -> Option<Self> {
+        Some(NP_UUID { value: [0u8; 16]})
+    }
+
     fn type_idx() -> (&'value str, NP_TypeKeys) { ("uuid", NP_TypeKeys::Uuid) }
     fn self_type_idx(&self) -> (&'value str, NP_TypeKeys) { ("uuid", NP_TypeKeys::Uuid) }
 
@@ -237,6 +245,14 @@ impl<'value> NP_Scalar<'value> for &NP_UUID {
 }
 
 impl<'value> NP_Value<'value> for &NP_UUID {
+
+    fn np_max_value<M: NP_Memory>(_cursor: &NP_Cursor, _memory: &M) -> Option<Self> {
+        None
+    }
+
+    fn np_min_value<M: NP_Memory>(_cursor: &NP_Cursor, _memory: &M) -> Option<Self> {
+        None
+    }
 
     fn type_idx() -> (&'value str, NP_TypeKeys) { NP_UUID::type_idx() }
     fn self_type_idx(&self) -> (&'value str, NP_TypeKeys) { NP_UUID::default().self_type_idx() }