+++ no_proto_rs/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.7.4 January 2, 2020
+- NP_Geo types no longer allow invalid values to be set into the buffer (outside lat/lng min & max values).
+- Added new `set_max` and `set_min` buffer methods to make it easer to make range query buffers.
+- Ran library through Miri, found some possible UB and fixed it.
+
 # 0.7.3 December 30, 2020
 - Added Prost to benchmarks.
 - Added documentation for zero-copy usage.